+++ dominion/cards/base_game/__init__.py
@@ -23,3 +23,5 @@ KINGDOM_CARDS = [
     VILLAGE,
     VILLAGE,
 ]
+
+STARTER_DECK = [ESTATE] * 3 + [COPPER] * 7
