+++ dominion/players/__init__.py
@@ -1,0 +1,2 @@
+from .deck import Deck
+from .players import HumanPlayer, ComputerPlayer
