+++ dominion/players/deck.py
@@ -1,18 +1,14 @@
-""" Player objects TODO: add docstring
+""" Deck TODO: docstring
 """
 
-from abc import ABC, abstractmethod
 import random  # For shuffle
 from collections import defaultdict
-from dominion.cards.base_game import ESTATE, COPPER
-from dominion.common import DeckPile, PlayerType, CardType
-from dominion.controller import Controller
-
-STARTER_DECK = [ESTATE] * 3 + [COPPER] * 7
+from dominion.common import DeckPile
+from dominion.cards.base_game import STARTER_DECK
 
 
 class Deck:
-    def __init__(self, starter_deck=STARTER_DECK):
+    def __init__(self, *, starter_deck=STARTER_DECK):
         self.draw_pile = starter_deck
         random.shuffle(self.draw_pile)
         self.discard_pile = []
@@ -102,85 +98,3 @@ class Deck:
             to_pile.insert(0, card)
 
         # TODO: allow int indexing as well as passing a list of names
-
-
-class Player(ABC):
-    def __init__(self, name):
-        self.name = name
-        self.deck = Deck()
-        self.modifiers = {}
-
-    @property
-    def hand(self):
-        """ Shorthand property for the player's hand."""
-        return self.deck.hand
-
-    @property
-    def treasure(self):
-        """ How much treasure the player has on hand."""
-        total = 0
-        for card in self.hand:
-            if card.kind == CardType.TREASURE:
-                total += card.value
-
-        # TODO: account for boons
-
-        if 'spent' in self.modifiers:
-            total -= self.modifiers['spent']
-
-        return total
-
-    def reset_modifiers(self):
-        self.modifiers = {}
-
-    @abstractmethod
-    def get_input(self, prompt, options, allow_skip=False):
-        pass
-
-
-class HumanPlayer(Player):
-    def __init__(self, name):
-        super().__init__(name)
-        self.type = PlayerType.HUMAN
-        self.controller = Controller()
-
-    def get_input(self, prompt, options, allow_skip=False):
-        """
-        @param options (dict): Number -> String option
-        """
-        return self.controller.get_input(prompt,
-                                         options,
-                                         allow_skip=allow_skip)
-
-    def show(self, text):
-        print(text)
-
-
-# Need to subclass the ComputerPlayer with an appropriate learn method and policy
-class ComputerPlayer(Player):
-    def __init__(self, name):
-        super().__init__(name)
-        self.type = PlayerType.COMPUTER
-
-    @abstractmethod
-    def reflect(self):
-        pass
-
-
-def get_playable_actions_as_options(cards):
-    """ Takes a list of cards and returns a dict from index to string option
-    """
-    options = {}
-    for idx, card in enumerate(cards):
-        if card.kind == CardType.ACTION:
-            options[idx] = card.name
-    return options
-
-
-def get_purchasable_cards_as_options(table, player):
-    options = {}
-    for idx, card in enumerate(table):
-        left = table[card]
-        if card.cost <= player.treasure and left > 0:
-            options[idx] = card.name
-    return options
